Change how InceptionBN switch its architecture.
In current implementation, we determine whether we use 1x1 convolution
and projection of the pooling path in __call__ by the existence of functions
(conv1 and poolp, respectively).

This commit changes these criteria so that they are identical to the ones
used in __init__ (out1 and proj_pool, respectively).

We can ensure these criteria are equivalent.

This change is for testability of InceptionBN.
I used MagicMock to replaces InceptionBN.f with the mocking object.
But MagicMock creates its attributes whenever they do not exist.
So in current implementation we cannot switch the architecture in unittests.