Fix incorrect dtype for ipiv buffer in cupy.linalg.inv
The ipiv parameter of sgetrf/dgetrf function is declared to be a
pointer to int array, but the current implementation allocates a
buffer ndarray with 'f' or 'd' dtype.

It is safe on all platforms supported by CUDA, because sizeof(int) <=
sizeof(float) on those platforms. But allocated buffer is larger than
necessary in the dgetrf case on those platforms, since sizeof(double)
is larger than sizeof(int).