Make `using_allocator` a private generator
As wrapping the Cython `using_allocator` with `contextmanager` is
problematic for documentation, make room for a Python wrapper that can
be a `contextmanager`. Do this by dropping the `contextmanager` bit from
the Cython implementation, mark it as private, and finally stop
importing it in `cupy.cuda`. Can follow this by wrapping it at the
Python level.