[WIP] some messy stuff for two-pass launch --- DO NOT USE!
It is clear that a two-pass launch is necessary to be close to the
performance of cub::DeviceReduce (for reducing whole array). However,
in order to do this we need to allocate temorary buffer of type_reduce.
This is not possible unless CuPy exposes its memory pool to C/C++
space so that we can call sizeof(type_reduce).