Use dtype objects instead of classes in the array API
The array API does not require any methods or behaviors on dtype objects,
other than that they be literals that can be compared for equality and passed
to dtype keywords in functions. Since dtype objects are already used by the
dtype attribute of ndarray, this makes it consistent, so that
func(dtype=<dtype>).dtype will give exactly <dtype> back, which will be the
same thing as numpy._array_api.<dtype>. This also fixes an issue in the array
API test suite due to the fact that dtype classes and objects are not equal as
dictionary keys.