Update the linear algebra functions in the array API namespace
For now, only the functions in from the main spec namespace are implemented.
The remaining linear algebra functions are part of an extension in the spec,
and will be implemented in a future pull request. This is because the linear
algebra functions are relatively complicated, so they will be easier to
review separately.

This also updates those functions that do remain for now to be more compliant
with the spec.