Restrict the array API namespace array operator type promotions
Only those type promotions that are required by the spec are allowed. In
particular, promotions across kinds, like integer + floating-point, are not
allowed, except for the case of Python scalars.

Tests are added for this.

This commit additionally makes the operators return NotImplemented on
unexpected input types rather than directly giving a TypeError. This is not
strictly required by the array API spec, but it is generally considered a best
practice for operator methods in Python.

This same thing will be implemented for the various functions in the array API
namespace in a later commit.