BUG: Fix new strides when array is both C and F-contiguous
In this case, NumPy will prefer the C contiguous strides, not the F
contiguous ones.

I will note that I do not think users should reuly on strides
if they rely on contiguity, and the old strides could be considered
just as valid as the new one.

Just as NumPy the strides of an F- and C-contiguous array created
as an F-contiguous one will have strides mutate on copy after this.
Currently, this happens for C contiguous ones.

Closes gh-5897