ENH: Add experimental string support to cupy
This is a squash of multiple changes (since it was unwieldy anyway)
and it is not a polished solution, rather basis for discussion.

To outline the design idea a little bit:

1. Create a C++ class to describes strings (this is very much
   a prototype).  This is in `numpystring.h`
2. Strings may need more casting work or work to define the correct
   result dtypes for example.  So, modelled on NumPy, I added a
   `resolve_dtypes` to the `_Op`.
   It is used by adding `_Op`s directly, see `_ufunc.py` or the `==`
   code.
3. Casting is a bit special, so I have a special path for it
   unfortunately, to encode that `int64 -> S` cast needs a certain
   length.
4. I added a refactor to auto-include the string header whenever
   the string is used (it may be added multiple times).
   (That is not strictly needed, but it allows not including it
   by default.  A similar thing could be done for complex now,
   but the change is not 100% as of now.)

There are some related smaller changes of course.  And e.g. the
way that loops gets selected work well enough for strings, but
has its limits (something that NumPy has a newer different mechanism
for because of that).  That is probably OK for the moment, I think.

Scalar strings do not work.  Casting is only implemented for strings and
for long<->string casts.  (All others will just fail compliation.)

---

Now, for me the main questions/discussion points would be about
whether this type of design seems fair, and it is easier to discuss
with some code :).
If it is, I will try to polish it up a bit, add scalars support and
at least better errors for the casts that do not work, I suspect.

This does add the complexity of `resolve_descriptors()` and it would choose
that complicated DTypes classes should have a C++ equivalent.  I don't
think that is strictly necessary, since the `routine` could be a function
so that it can be adapted for the string length in theory!
(but it probably helps if at least conversions work for the moment.)

I am not sure strings are worth that complexity!  OTOH, I have hope that
we can create a C++ tuple-like class to describe a large class of structured
DTypes (or wait for CCCL and use tuples, the main thing to work out is that in
NumPy `arr = np.ones(3, "i,i,i")[["f0", "f1"]]` creates a dtype with a "hole"
and that hole needs to be encoded, but I think we can do that with a simple
no-op C++ object!).

It might also be interesting for the longer-term future, to help set up
infrastructure that might allow users to write custom kernels for custom
C++ classes (which of course would need some form of "dtype" object either
compatible with NumPy, or CuPy would have to wrap NumPy)?