BUG: remove short int loops from true_divide
The issue this: consider `cupy.uint8(2) / (-2)`. Since -2 is a
python scalar, the ufunc selects a 'BB->d' loop based on the type of
the other argument. Then the ufunc kernel declared
`in1_type in1` which is initialized `uint8_t in1(-2)`.

With only 'int64, int64 -> float64' and 'uint64, uint64_5 -> float64'
loops (this PR), the problem appears only for uint64(...) / (-2),
and that's probably rare enough.