Fix copyto for NumPy 2 compatibility
NumPy two has some bad definitions with `copyto` and NEP 50
in that `copyto` currently ignores it. That is a trickier NumPy problem.

However, it is very hard to write code that does the right thing on
both versions and even harder to do the same thing as NumPy 2 (because it
is not quite consistent).

Short of a version switch, the only good way seems to be to just call
`copyto` directly for Python scalar input.

Closes gh-8408, gh-8391