Use custom less instead of specializing thrust
The current way of specializing thrust::less::operator() is uncommon in C++ and not forward compatible to recent and future changes to CCCL. The proper way is to use a custom comparison object with the desired behavior.

For sorting floating-points, thrust::less is used directly, since then thrust uses radix sort, which sorts NaNs to the back.