upfirdn: mode=None -> mode="constant"
scipy.signal.upfirdn supports a number of modes, of which the default is
"constant". cupyx supports only this mode, but used `None` to select it.
Code that didn't pass the argument would work interchangeably between
scipy and cupyx, but code that explicitly passed `mode="constant"` would
raise an exception under cupyx.

With this change, cupyx behaves the same as scipy for this mode: not
passing the argument, or explicitly passing `mode="constant"` works, and
explicit passing `mode=None` raises an exception. Note that the latter
is a **backwards-incompatible** change.

Closes #8449.