Extend runtime header search logic to conda
As CuPy also needs a bit of help searching for headers in the conda
case, extend the logic used to search for wheel headers to also search
for conda headers. In particular this searches the targets directory for
the platform where CuPy is being run on so that it can find headers
there.