Set the archiver `ar`
Distutils and Conda both set this. Even though we don't use it now, it
could be used in the future. Also it doesn't look like Distutils handles
it gracefully if this is unset. So configure to `ar` for the build
platform (not host) as is consistent with the rest of these.