Speed up dlpack export and ignore stream for CPU requests
The stream ignoring is just because not ignoring it came from me
forgetting my own argument:  The stream should be tied to the requested
device and the requested device has no stream defined.
(Because of that, the old way was safe but I doubt there is much need
for it, so I think we should simplify it and not give others the wrong idea.)

Otherwise some micro optimizations for certain degrees of micro.  Not
all may be wanted, so happy to revert some if that feels better:
* Coalescing allocation and using PyMem* probably doesn't do much
  but also doesn't make things worse to read to me.
* The new `pointerGetMemoryType` is likely the most unclear.  Avoids
  an object creation and attribute lookup (probably around 40ns).
* The type-number list avoids one attribute lookup and the conversion,
  probably around 30ns, so simple structure but maybe rather verbose.

The biggest single thing is just the loop variable `n` I think,
which Cython didn't auto-optimize (probably because of the second
loop using the same variable?).

All of this got me from around 900ns to around 350ns, almost 100(?)ns
of that may just be the `cudaPointerGetAttributes` so I suspect there
isn't too much left to achieve here.
(The cuda stream/device handling just adds some complexity in the end.)
(Note timings were done with, now split out thread-local refactor.)

Just a note, though.  When streams are used, the stream order logic will
add almost 4us for me, meaning that all of this is unimportant...