Replace fastrlock with C++ recursive_mutex
It seems that fastrlock tends to be a problem with new Python versions
and also freethreading.
This proposes to replace it with a C++ mutex, making it recursive/re-entrant
just because I am not 100% sure about things.

I am not sure if there are very subtle issues (subinterpreters?)
if using a C++ mutex, but I don't think those are of a big concern.
FWIW, NumPy also uses a lot of C++ mutexes for free-threading support at least.

Since Python 3.10 has direct GC C-API functions in the stable API just added
those, I didn't time if it is relevant, but it seemed just as well/nice.