ENH: Accept nested CAI arrays in cupy.asarray and indexing
The indexig path went through `min_scalar_type`, I am not sure
that makes sense, but kept it for now.
However, that required a slight change there because the old
call expected a list/tuple (and this is a bare CAI), but full
conversion seemed not helpful.

I changed the flattening code to append recursively. It is still
depth first.
The reason is that it is now necessary to convert to a cupy array
earlier or `obj.ndim` failed in the tests for CAI objects.

Note: That is a very subtle change, because that path requires a
`.shape` and `.dtype` attribute for objects supporting
`__cupy_get_ndarray__`.
If that is a worry, we should tweak this a bit.