MAINT: Specify texture address precisely
I tried using `profile=True` and I think I also ran into this once
before (not sure due to what).
In some cases Cython generates code to expose this to Python and
then it fails casting to `int*` when compiling the generated C++
code.

Maybe whether it fails compiling is dependent on more things, since
it doesn't happen in CI?