[bug] Include type_traits in filters
test_filters.py relies on the type_traits library. When JIT
compiling code that uses something `std::enable_if` in these
tests, users are met with a compilation error saying that
`std::enable_if` can't be found and to try using the namespace
`__hip_internal::enable_if` instead. This is likely because
`std::enable_if` was previously defined in ROCm versions < 7,
but is no longer the case.

When printing the failing generated code, it became clear that
the type_traits header was not included, which caused the compilation
failure.

This is easily fixed by including `type_traits` in the filters
JIT strings.