BUG,PERF: Fixup and simplify GUFuncs to also speed them up
This reduces the GUfunc overhead of a simple matmul from ~60% of the
operation to probably 10%, I expect around 1/3 of the time if the
matmul core itself is well optimized.

Mainly, also fixes a bunch of bugs, unfortunately... this is rather
complex code so it is hard to review.
(I am sure it fixes a lot of bugs more than it opens, but it's hard
to be sure everything is covered...)

It does have one larger change (also a bug fix) that I am aware of:
The old code incorrectly broadcast _core_ dimensions, this is not
allowed by NumPy unless the dimension is specified as `n|1`.
For matmul this just changes the error if someone actually relied
on it, they would have to switch to `|1` on newer versions.