Add fast-path for gufunc (specifically matmul)
Since simple `cupy.matmul` calls, half of the overhead is still
inside the gufuncs even with the previous PR, this suggests adding
a fast-path mechanism.

I would love to avoid the fast-path here, but I doubt it is
worthwhile to carefully Cythonize the gufuncs and a 2-3x speeup
is just still a bit underwhelming.
(Fast-paths are always an opportunity for bugs after all...)