File cache: use os.replace (clarity) and accept PermissionError
This uses `os.replace()` since it is a bit more clear that a within-folder
move should be atomic.
It also allows `PermissionError` on windows since it is possible that the
file was already created by another thread and opened for use and windows
will then raise a `PermissionError` when the replace fails.

Unfortunately, I am not sure there is a good way to test this.

Closes gh-9463