BUG: Fix typo in advise and prefetch
The struct is already the location, so this added one level.
Strangely, Cython chooses to not fail compilation but rather raise
a runtime error (pretending that this is a dict).

Added a very simple test, but happy to iterate on it if it doesn't
seem ideal.

Just as a note: this API is documented as experimental.