PERF: Lower thread handling into Cython
Thread local lookups are not very slow but do add up to a noticable
amount. This lowers them into C.

The tricky part here is garbage collection when the Python thread is
destructured. This works roughly as it does for thread-local also:
https://github.com/python/cpython/blob/9f2a34af747e2fc95f3de8a3bb5e1d7ac10d6d04/Lib/_threading_local.py#L62

I.e. we create a weakref on the thread object with a destructor that
cleans up our thread-state.
The second trick here is that this is done via a bound method and this
only works on account of a reference cycle, so that the `@cython.no_gc`
decorator is actually required to make it work correctly.
(Otherwise, we would have to store it elsewhere.)

In many places this doesn't matter, but I wanted to align things.
I re-structured the "handles" also because the in-line import is not
great.