TST: Skip many tests when running with pytest-run-parallel
This skips many tests when running in free-threading mostly.
It also does some smaller re-organizations at times.

The random tests (at least repeats) are judiciously skipped, it is
just a bit of a tricky setup and there should be quite a lot of
tests that still run anyway.

One change that may be larger is setting warnings to "always" which
seemed necessary for some warnings tests, if that leads to a lot of
spam in CI, we may want to skip that for now.