DEV: Make `CUPY_TEST_GPU_LIMIT` more reliable.
Not that `CUPY_TEST_GPU_LIMIT` is a game-changer probably, but it
didn't actually work reliably (I assume it used to work).
I am not sure why exactly, it may be that the fixture is too late
because it is run after test collection (which then may e.g.
create cupy arrays).

This pattern using the environment variables was probably not available
when it was introduced initially, but works now.

(N.B. at least for high-memory tests using two runners on the same GPU
probably doesn't work, xdist groupings may be a way to solve this.)