DEV: Fix `CUPY_TEST_GPU_LIMIT`
Not that `CUPY_TEST_GPU_LIMIT` is a game-changer probably, but it
didn't actually work (I assume it used to work), due to cupy being
imported.
A fixture seems too late, as it is run after tests are gathered.

Also the exception imports from `cupy.exceptions` were enough to break
things, but we can get them from NumPy again now.