Make FFT config more thread-safe and adatapt FFT tests
This contains two sets of changes around thread-safety and FFTs.

First, it modifies `cupy.fft.config` to be an object rather than a module
to allow making `.devices` thread-local, this may be up for discussion
but for the FFT config it seemed sensible to me.

I decided to use contextvars for this, which I think is the best pattern
(thread and context safe), but they remain private for now.

The tests also need fairly extensive changes anyway to make them thread-safe,
one with custom decoartors (fixtures are not run for each thread in
`pytest-run-parallel`).
They were also digging into the FFT internals so needed to adapt to that.

(My local runs had some problems with the callback tests with
or without threading, so this may not be 100%.)