MAINT: Remove `-fno-gnu-unique` again
I was using sanitizers and tried compiling with clang, which did not
like `-fno-gnu-unique`.

Based on the history, this seems to have been to avoid symbol clashes
via cub/thrust. Based on feedback from the CCCL team, CCCL now makes
sure to use a namespace that is system and version dependent, so
I think this option should be unnecessary now.