MAINT: Use `__builtin_bit_cast` or union for half-infinity constexpr
GCC seems to accept the reinterpret cast, but clang refuses it as it
is technically not a constexpr.
The nicest thing would be `std::bit_cast`, but until we can, use
`__builtin_bit_cast` on compilers we know are OK and otherwise
use `union` which should be fine, but may technically not be C++
standard conform.