MAINT: Use only union for half-infinity constexpr
GCC seems to accept the reinterpret cast, but clang refuses it as it
is technically not a constexpr.
The nicest thing would be `std::bit_cast`, but while I am not sure that
the union is the best solution, it seems like a good enough solution
that should work without the need for `#ifdefs` (I don't see a reason
why the `#ifdef` would be helpful).